model/requirement.txt
numpy
pandas
scikit-learn
tensorflow
pymongo
python-dotenv